sw_services:xilsecure:Fix for MISRA-C Rule 10.4 violation
MISRA-C Rule 10.4: Essential type of the left hand operand is not the same as that of the right operand
Solution: Typecast accordingly


Acked-for-series: Durga Challa <durga.challa@amd.com>

+++ lib/sw_services/xilsecure/src/server/core/sha/sha_pmxc/xsecure_sha_ipihandler.c
@@ -78,7 +78,7 @@ int XSecure_ShaIpiHandler(XPlmi_Cmd *Cmd)
 
 	ApiId = Cmd->CmdId & XSECURE_API_ID_MASK;
 
-	if (ApiId == XSECURE_API_SHA2_OPERATION) {
+	if (ApiId == (u32)XSECURE_API_SHA2_OPERATION) {
 		XSecureShaInstPtr = XSecure_GetSha2Instance(XSECURE_SHA_1_DEVICE_ID);
 		Core = XPLMI_SHA2_CORE;
 	}
